requirements.txt
ultralytics
opencv-python
numpy
pandas
pyyaml
mlflow
boto3
fastapi
uvicorn
python-multipart
fastapi